requirements.txt
streamlit>=1.35.0
gspread>=6.0.0
google-auth>=2.29.0
pandas>=2.0.0
numpy>=1.26.0
reportlab>=4.2.0
